KeyBinding cleanup
When a key is rebound, it's previous usage in KeyConfigPane is now properly cleaned. KeyMap has also been fixed so that there can't be duplicate values.

+++ src/tunnelers/Game/IO/AControlScheme.java
@@ -8,16 +8,16 @@ import javafx.scene.input.KeyCode;
  */
 public abstract class AControlScheme {
 
-	private static int controlSchemesCreated = 0;
+	private static byte controlSchemesCreated = 0;
 
-	protected final int schemeID;
+	protected final byte schemeID;
 	protected int playerID;
 
 	public AControlScheme() {
 		schemeID = controlSchemesCreated++;
 	}
 
-	public int getID() {
+	public byte getID() {
 		return this.schemeID;
 	}
